clip/requirements.txt
filelock==3.16.1
fsspec==2024.10.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
pillow==11.0.0
sympy==1.13.1
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0
typing_extensions==4.12.2
